Move various code into a new lib/ directory
