+++ lib/watchdog.js
